creating container component (#31)
- Made the panel a child of the modal, so it can be used in apps without the modal wrapper.
- Created a ExportVideoPanelContainer component for states and recording setup.
- Rough fix for zoom out.
- Removed some redundant react updates
- MIT Licenses
