+++ examples/kepler-integration/src/components/export-video/export-video-panel.js
@@ -18,63 +18,17 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-import React, {Component} from 'react';
+import React from 'react';
 import styled, {withTheme} from 'styled-components';
-import {Button, Icons} from 'kepler.gl/components';
-
-import ExportVideoPanelSettings from './export-video-panel-settings';
-import ExportVideoPanelFooter from './export-video-panel-footer';
-import {ExportVideoPanelPreview} from './export-video-panel-preview'; // Not yet part of standard library. TODO when updated
-import {parseSetCameraType} from './parse-set-camera-type';
-import {DEFAULT_PADDING, DEFAULT_ICON_BUTTON_HEIGHT} from './constants';
-
-import {easing} from 'popmotion';
-
-import {
-  DeckAdapter,
-  DeckScene,
-  CameraKeyframes,
-  WebMEncoder,
-  JPEGSequenceEncoder,
-  PNGSequenceEncoder,
-  PreviewEncoder,
-  GifEncoder
-} from '@hubble.gl/core';
-
-// import {DEFAULT_TIME_FORMAT} from 'kepler.gl';
-// import moment from 'moment';
-import {messages} from 'kepler.gl/localization';
 import {IntlProvider} from 'react-intl';
 
-// function setFileNameDeckAdapter(name) {
-//   encoderSettings.filename = `${name} ${moment()
-//     .format(DEFAULT_TIME_FORMAT)
-//     .toString()}`;
-// }
+import {messages} from 'kepler.gl/localization';
+import {Button, Icons} from 'kepler.gl/components';
 
-/* function setResolution(resolution){
-  if(resolution === 'Good (540p)'){
-    adapter.scene.width = 960;
-    adapter.scene.height = 540;
-  }else if(resolution === 'High (720p)'){
-    adapter.scene.width = 1280;
-    adapter.scene.height = 720;
-  }else if(resolution === 'Highest (1080p)'){
-    adapter.scene.width = 1920;
-    adapter.scene.height = 1080;
-  }
-}*/
-
-// TODO:
-
-// Changes Timestamp function
-// Camera function (preset keyframes) DONE
-// File Name function DONE
-// MediaType function DONE
-// Quality function
-// Set Duration function
-// Calculate File Size function
-// Render Function DONE
+import {DEFAULT_PADDING, DEFAULT_ICON_BUTTON_HEIGHT} from './constants';
+import ExportVideoPanelSettings from './export-video-panel-settings';
+import {ExportVideoPanelPreview} from './export-video-panel-preview'; // Not yet part of standard library. TODO when updated
+import ExportVideoPanelFooter from './export-video-panel-footer';
 
 const IconButton = styled(Button)`
   padding: 0;
@@ -140,171 +94,52 @@ const PanelBody = ({
 );
 
 const Panel = styled.div`
-  width: ${props => props.settingsWidth}px;
+  width: ${props => props.exportVideoWidth}px;
 `;
 
-class ExportVideoPanel extends Component {
-  static defaultProps = {
-    settingsWidth: 980
-  };
-
-  constructor(props) {
-    super(props);
-
-    this.setMediaTypeState = this.setMediaTypeState.bind(this);
-    this.setCameraPreset = this.setCameraPreset.bind(this);
-    this.setFileName = this.setFileName.bind(this);
-    // this.setQuality = this.setQuality.bind(this);
-    this.getCameraKeyframes = this.getCameraKeyframes.bind(this);
-    this.getDeckScene = this.getDeckScene.bind(this);
-    this.onPreviewVideo = this.onPreviewVideo.bind(this);
-    this.onRenderVideo = this.onRenderVideo.bind(this);
-
-    this.state = {
-      mediaType: 'GIF',
-      cameraPreset: 'None',
-      fileName: 'Video Name',
-      //  quality: "High (720p)",
-      viewState: this.props.mapData.mapState,
-      durationMs: 1000,
-      encoderSettings: {
-        framerate: 30,
-        webm: {
-          quality: 0.8
-        },
-        jpeg: {
-          quality: 0.8
-        },
-        gif: {
-          sampleInterval: 1000
-        },
-        filename: 'kepler.gl'
-      },
-      adapter: new DeckAdapter(this.getDeckScene)
-    };
-  }
-
-  getCameraKeyframes(prevCamera = undefined) {
-    const {viewState, cameraPreset, durationMs} = this.state;
-
-    return new CameraKeyframes({
-      timings: [0, durationMs],
-      keyframes: [
-        {
-          longitude: viewState.longitude,
-          latitude: viewState.latitude,
-          zoom: viewState.zoom,
-          pitch: viewState.pitch,
-          bearing: viewState.bearing
-        },
-        parseSetCameraType(cameraPreset, viewState)
-      ],
-      easings: [easing.easeInOut]
-    });
-  }
-
-  getDeckScene(animationLoop) {
-    const keyframes = {
-      camera: this.getCameraKeyframes()
-    };
-    const currentCamera = animationLoop.timeline.attachAnimation(keyframes.camera);
-
-    return new DeckScene({
-      animationLoop,
-      keyframes,
-      lengthMs: this.state.durationMs, // TODO change to 5000 later. 1000 for dev testing
-      width: 480,
-      height: 460,
-      currentCamera
-    });
-  }
-
-  setMediaTypeState(media) {
-    this.setState({
-      mediaType: media
-    });
-  }
-  setCameraPreset(option) {
-    this.setState({
-      cameraPreset: option
-    });
-  }
-  setFileName(name) {
-    this.setState({
-      fileName: name.target.value
-    });
-    // setFileNameDeckAdapter(name.target.value);
-  }
-  /* setQuality(resolution){
-    this.setState({
-      quality: resolution
-    });
-    setResolution(resolution);
-  }*/
-
-  onPreviewVideo() {
-    const {adapter, encoderSettings} = this.state;
-
-    const onStop = () => {};
-    adapter.render(PreviewEncoder, encoderSettings, onStop, this.getCameraKeyframes);
-  }
-
-  onRenderVideo() {
-    const {adapter, encoderSettings, mediaType} = this.state;
-    let Encoder = PreviewEncoder;
-    const onStop = () => {};
-
-    if (mediaType === 'WebM Video') {
-      Encoder = WebMEncoder;
-    } else if (mediaType === 'PNG Sequence') {
-      Encoder = PNGSequenceEncoder;
-    } else if (mediaType === 'JPEG Sequence') {
-      Encoder = JPEGSequenceEncoder;
-    } else if (mediaType === 'GIF') {
-      Encoder = GifEncoder;
-    }
-
-    adapter.render(Encoder, encoderSettings, onStop, this.getCameraKeyframes);
-  }
-
-  render() {
-    const {settingsWidth, handleClose, mapData} = this.props;
-    const settingsData = {
-      mediaType: this.state.mediaType,
-      cameraPreset: this.state.cameraPreset,
-      fileName: this.state.fileName,
-      resolution: this.state.quality
-    };
-
-    const {exportSettings, adapter} = this.state;
-
-    return (
-      <IntlProvider locale="en" messages={messages.en}>
-        <Panel settingsWidth={settingsWidth} className="export-video-panel">
-          <PanelClose handleClose={handleClose} />
-          <StyledTitle className="export-video-panel__title">Export Video</StyledTitle>
-          <PanelBody
-            mapData={mapData}
-            exportSettings={exportSettings}
-            adapter={adapter}
-            setMediaType={this.setMediaTypeState}
-            setCameraPreset={this.setCameraPreset}
-            setFileName={this.setFileName}
-            //  setQuality={this.setQuality}
-            settingsData={settingsData}
-            setViewState={viewState => {
-              this.setState({viewState});
-            }}
-          />
-          <ExportVideoPanelFooter
-            handleClose={handleClose}
-            handlePreviewVideo={this.onPreviewVideo}
-            handleRenderVideo={this.onRenderVideo}
-          />
-        </Panel>
-      </IntlProvider>
-    );
-  }
-}
+const ExportVideoPanel = ({
+  // UI Props
+  exportVideoWidth,
+  handleClose,
+  // Map Props
+  mapData,
+  setViewState,
+  // Settings Props
+  settingsData,
+  setMediaTypeState,
+  setCameraPreset,
+  setFileName,
+  // setQuality,
+  // Hubble Props
+  exportSettings,
+  adapter,
+  handlePreviewVideo,
+  handleRenderVideo
+}) => {
+  return (
+    <IntlProvider locale="en" messages={messages.en}>
+      <Panel exportVideoWidth={exportVideoWidth} className="export-video-panel">
+        <PanelClose handleClose={handleClose} />
+        <StyledTitle className="export-video-panel__title">Export Video</StyledTitle>
+        <PanelBody
+          mapData={mapData}
+          exportSettings={exportSettings}
+          adapter={adapter}
+          setMediaType={setMediaTypeState}
+          setCameraPreset={setCameraPreset}
+          setFileName={setFileName}
+          //  setQuality={this.setQuality}
+          settingsData={settingsData}
+          setViewState={setViewState}
+        />
+        <ExportVideoPanelFooter
+          handleClose={handleClose}
+          handlePreviewVideo={handlePreviewVideo}
+          handleRenderVideo={handleRenderVideo}
+        />
+      </Panel>
+    </IntlProvider>
+  );
+};
 
 export default withTheme(ExportVideoPanel);
