+++ examples/kepler-integration/src/components/export-video/parse-set-camera-type.js
@@ -1,3 +1,23 @@
+// Copyright (c) 2020 Uber Technologies, Inc.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
 import {point} from '@turf/helpers';
 import transformTranslate from '@turf/transform-translate';
 
@@ -48,9 +68,9 @@ export function parseSetCameraType(strCameraType, viewState) {
 
   if (match[0] === 'Zoom') {
     if (match[1] === 'In') {
-      modifiedViewState.zoom = 15;
+      modifiedViewState.zoom = modifiedViewState.zoom + 3;
     } else if (match[1] === 'Out') {
-      modifiedViewState.zoom = 15;
+      modifiedViewState.zoom = modifiedViewState.zoom - 3;
     }
   }
   return modifiedViewState;
